@@ -1,18 +1,18 @@
 <template>
   <div id="app">
-    <img alt="Vue logo" src="./assets/logo.png" />
-    <HelloWorld msg="Welcome to Your Vue.js + TypeScript App" />
+    <TComp />
   </div>
 </template>
 
 <script lang="ts">
-import Vue from "vue";
-import HelloWorld from "./components/HelloWorld.vue";
+import { defineComponent } from "vue";
 
-export default Vue.extend({
+import TComp from "./components/TComp.vue";
+
+export default defineComponent({
   name: "App",
   components: {
-    HelloWorld,
+    TComp,
   },
 });
 </script>
